18 integrate docker compose on azure vm (#38)
## Description
This pull request handles the MQTT Broker which handles MQTT messages to
and from client instances. The received messages are re-routed to the
database through the MQTT_bridge.py. The app.py contains a draft of an
API, and handles the communication between broker and database.
Additionally we have included script files to ensure easy installation
and restart of the docker images to be used in virtual machines with
Azure.

## Implementation details

- The mqtt_bridge.py handles the MQTT connection through an MQTT broker
and a corresponding client. Ultimately the messages recieved at the
client is re-routed to our timescale database. The bridge ensures TLS
communication between the MQTT clients.
- The app.py, is currently a draft API with endpoints, although these
are only accessible on the virtual machine with localhost. These are
supposed to be exposed through HTTPS in the future, this will be done
through issue #19.
- The entire setup is dockerized within a docker virtual network,
allowing us to easily start all containers.
- Scripts have been made to ensure easy installation, restart and
cleanup on virtual machines.
- Install.sh will installs the necessary dependencies and correct
configurations. Additionally, it creates a new certificate used for TLS
communication.
- Restart.sh will restart the MQTT/TLS service, prompting the user if a
new certificate should be created. Database volume is retained.
- Cleanup.sh will take down the MQTT/TLS service, prompting the user if
the certificates as well as the database volume should be retained.

---------

Co-authored-by: martivj <martinvagseterjakobsen@yahoo.no>