37 make a simple frontend for sensor data (#55)
### Backend

There is now a functioning .NET backend that processes data from the
broker API in specialized endpoints. The swagger documentation and
README shows how to interact with the .NET API and how to build and run
the dotnet instance. Unit tests have also been added, to ensure expected
behavior when mocking the broker API.

### Frontend

There is now a Nuxt/Vue.js frontend that uses the new .NET backend to
display sensor data. It supports 4 tabs in a single page application for
the sensor data and configuration options with the chip setup. The
README describes what each component does, and how to run, build and
preview the application.

Closes the following issues:
Close #36
Close #37

---------

Co-authored-by: Magnus Hansen Åsheim <magnhaas@stud.ntnu.no>
Co-authored-by: Magnus Åsheim <magnus-h@hotmail.no>
Co-authored-by: leifeggenfellner <eggenfellner@protonmail.com>
Co-authored-by: Sigve Kulsrud <sigvefk@ntnu.no>